Ajustes de estabilizacion

+++ src/main/java/com/drr/ucompensar/repository/UsuarioRepository.java
@@ -3,6 +3,6 @@ package com.drr.ucompensar.repository;
 import com.drr.ucompensar.core.GenericRepository;
 import com.drr.ucompensar.entity.UsuarioEntity;
 
-public interface UsuarioRepository extends GenericRepository<UsuarioEntity, Integer> {
+public interface UsuarioRepository extends GenericRepository<UsuarioEntity, Long> {
 	
 }
